discord/requirements.txt
pip
py-cord
python-dotenv
Flask
requests
python-whois
cryptography
pyOpenSSL
gunicorn